# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Premium Tools (AIPT) ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Categories ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, price_bdt, category_id, is_active, is_featured, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price_bdt:
          type: number
        original_price_bdt:
          type: number
        category_id:
          type: integer
        category_name:
          type: string
        image_url:
          type: string
        features:
          type: array
          items:
            type: string
        duration_days:
          type: integer
        is_active:
          type: boolean
        is_featured:
          type: boolean
        stock_count:
          type: integer
        order_count:
          type: integer
        created_at:
          type: string
    CreateProductBody:
      type: object
      required: [name, price_bdt, category_id]
      properties:
        name:
          type: string
        description:
          type: string
        price_bdt:
          type: number
        original_price_bdt:
          type: number
        category_id:
          type: integer
        image_url:
          type: string
        features:
          type: array
          items:
            type: string
        duration_days:
          type: integer
        is_active:
          type: boolean
        is_featured:
          type: boolean
        stock_count:
          type: integer

    # ── Category ─────────────────────────────────────────────────────────────
...